Move ready listener to own file
