+++ src/events/events.ts
@@ -44,40 +44,6 @@ export type EventModule = {
 export function registerEventHandlers({ nessie, mixpanel }: Props) {
   //------
   /**
-   * Event handler that fires once when nessie boots up and succesfully logs in
-   */
-  nessie.once('ready', async () => {
-    await registerApplicationCommands();
-    try {
-      const testChannel = nessie.channels.cache.get('889212328539725824');
-      const logChannel = nessie.channels.cache.get('899620845436141609');
-      testChannel && testChannel.isText() && testChannel.send("I'm booting up! (◕ᴗ◕✿)"); //Sends to test bot channel in nessie's canyon
-      /**
-       * Initialise Database and its tables
-       * Will create them if they don't exist
-       * See relevant files under database/* for more information
-       */
-      createGuildTable(nessie.guilds.cache, nessie);
-      createStatusTable();
-      /**
-       * Changes Nessie's activity when the current map has switched over to the next
-       * Refer to the setCurrentMapStatus function for more information
-       */
-      const brPubsData = await getBattleRoyalePubs(); //Get data of br map rotation
-      nessie.user && nessie.user.setActivity(brPubsData.current.map); //Set current br map as activity status
-      sendHealthLog(brPubsData, logChannel, true); //For logging purpose
-      setCurrentMapStatus(brPubsData, logChannel, nessie); //Calls status display function
-      // const statusScheduler = initialiseStatusScheduler(nessie); //Initialises auto status scheduler
-      // statusScheduler.start(); //Starts the scheduler
-
-      const statusSchedule = scheduleStatus(nessie);
-      statusSchedule.start();
-    } catch (e) {
-      console.log(e); //Add proper error handling
-    }
-  });
-  //------
-  /**
    * Event handlers for when nessie is invited to a new server and when he is kicked. Will be opting out of guild update as I don't really need to do anything with that
    * Sends notification to channel in Nessie's Canyon
    * guildCreate - called when Nessie is invited to a server
@@ -214,90 +180,3 @@ export function registerEventHandlers({ nessie, mixpanel }: Props) {
     await sendErrorLog({ nessie, error, type, uuid, ping: true });
   });
 }
-
-//TODO: Maybe move these functions in their separate files at some point
-/**
- * In charge of correctly displaying current battle royale pubs rotation in nessie's activity status
- * As the maps have varying durations, needed to figure out a way to dynamically change the timeout after each call
- * Accomplished this by creating a intervalRequest function that has a setTimeout that calls itself as its callback
- * Inside the interval function we can then properly get the current timer and update accordingly
- */
-const setCurrentMapStatus = (data: any, channel: any, nessie: any) => {
-  const fiveSecondsBuffer = 5000;
-  let currentBrPubsData = data;
-  let currentTimer = data.current.remainingSecs * 1000 + fiveSecondsBuffer;
-  const intervalRequest = async () => {
-    try {
-      const updatedBrPubsData = await getBattleRoyalePubs();
-      /**
-       * Checks to see if the data taken from API is accurate
-       * Was brought to my attention that the status was displaying the wrong map at one point
-       * Not sure why this is happening so just adding a notification when this happens again
-       * Don't really want to add extra code for now, if it happens again then i'll fix it
-       */
-      const isAccurate = currentBrPubsData.next.code === updatedBrPubsData.current.code;
-      currentBrPubsData = updatedBrPubsData;
-      currentTimer = updatedBrPubsData.current.remainingSecs * 1000 + fiveSecondsBuffer;
-      nessie.user.setActivity(updatedBrPubsData.current.map);
-      sendHealthLog(updatedBrPubsData, channel, isAccurate);
-      setTimeout(intervalRequest, currentTimer);
-    } catch (e) {
-      console.log(e);
-      channel.send('<@183444648360935424> WHOOPS SOMETHING WENT WRONG');
-    }
-  };
-  setTimeout(intervalRequest, currentTimer); //Start initial timer
-};
-/**
- * Function to register application commands
- * Application commands are different from regular prefix commands as instead of the bot directly responding to messages, discord would be the one doing it
- * In order to do that, we have to register the bot's application commands to discord first before it can be used
- * Two types of application commands:
- * - Guild Commands
- * - Global Commands
- * Basicailly guild commands can only be used in that server it was registered in
- * While global commands can be used to every server that bot is in
- * Main difference between the two apart from server constraints are that app commands are instantly registered in guilds while global would take up to an hour for changes to appear
- */
-const registerApplicationCommands = async () => {
-  const isInDevelopment = ENV === 'dev';
-  const publicCommandList = Object.keys(appCommands)
-    .map((key) => !appCommands[key].isAdmin && appCommands[key].data)
-    .filter((command) => command)
-    .map((command) => command.toJSON());
-  const adminCommandList = Object.keys(appCommands)
-    .map((key) => appCommands[key].isAdmin && appCommands[key].data)
-    .filter((command) => command)
-    .map((command) => command.toJSON());
-  const fullCommandList = Object.keys(appCommands)
-    .map((key) => appCommands[key].data)
-    .filter((command) => command)
-    .map((command) => command.toJSON());
-
-  const rest = new REST({ version: '9' }).setToken(BOT_TOKEN);
-
-  if (isInDevelopment) {
-    //Guild register
-    try {
-      await rest.put(Routes.applicationGuildCommands('929421200797626388', GUILD_ID), {
-        body: fullCommandList,
-      });
-      console.log('Successfully registered guild application commands');
-    } catch (e) {
-      console.log(e);
-    }
-  } else {
-    //Global Register
-    //TODO: Maybe create a script one day to delete global commands for test bot
-    //TODO: Make fetching of bot id dynamic as it will either use production or testing id
-    try {
-      await rest.put(Routes.applicationCommands('889135055430111252'), { body: publicCommandList });
-      await rest.put(Routes.applicationGuildCommands('889135055430111252', GUILD_ID), {
-        body: adminCommandList,
-      });
-      console.log('Successfully registered global application commands');
-    } catch (e) {
-      console.log(e);
-    }
-  }
-};
