+++ src/services/adapters.ts
@@ -5,49 +5,49 @@ import { ALS_API_KEY } from '../config/environment';
 const url = `https://api.mozambiquehe.re/maprotation?version=2&auth=${ALS_API_KEY}`;
 
 //TODO: Add schema for ALS responses
-exports.getBattleRoyalePubs = async () => {
+export const getBattleRoyalePubs = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data.battle_royale;
   });
   console.log(data);
   return data;
 };
-exports.getBattleRoyaleRanked = async () => {
+export const getBattleRoyaleRanked = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data.ranked;
   });
   console.log(data);
   return data;
 };
-exports.getArenasPubs = async () => {
+export const getArenasPubs = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data.arenas;
   });
   console.log(data);
   return data;
 };
-exports.getArenasRanked = async () => {
+export const getArenasRanked = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data.arenasRanked;
   });
   console.log(data);
   return data;
 };
-exports.getControlPubs = async () => {
+export const getControlPubs = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data.control;
   });
   console.log(data);
   return data;
 };
-exports.getLimitedTimeEvent = async () => {
+export const getLimitedTimeEvent = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data.ltm;
   });
   console.log(data);
   return data;
 };
-exports.getRotationData = async () => {
+export const getRotationData = async () => {
   const data = await axios.get(url).then((response: any) => {
     return response.data;
   });
